# public/assets/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kiracloud API",
    "version": "v2.0",
    "description": "Simple and powerful REST API — downloaders, search tools, and system utilities."
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/api/downloader/youtube": {
      "get": {
        "summary": "YouTube Downloader",
        "description": "Download high-quality MP4 video and 320kbps MP3 audio from any YouTube URL.",
        "tags": ["Downloader"],
        "parameters": [
          { "name": "url", "in": "query", "required": true, "schema": { "type": "string", "example": "https://youtu.be/PPKHWvVAN9U" } },
          { "name": "resolusi", "in": "query", "required": false, "schema": { "type": "string", "example": "max", "default": "max" } }
        ],
        "responses": {
          "200": { "description": "Video and audio download links returned." },
          "500": { "description": "Internal server error." }
        }
      }
    },
...
  }
}